refactoring auth model to bloc
