+++ lib/domain/blocs/auth_view_cubit.dart
@@ -1,0 +1,10 @@
+class AuthViewCubitState {
+  final String errorMessage;
+  final bool _isAuthProgress;
+
+  AuthViewCubitState(this.errorMessage, bool isAuthProgress) : _isAuthProgress = isAuthProgress;
+
+  bool get canStartAuth => !_isAuthProgress;
+
+  bool get isAuthProgress => _isAuthProgress;
+}
