+++ lib/ui/widgets/loader_widget/loader_widget.dart
@@ -1,6 +1,6 @@
 import 'package:comics_db_app/ui/components/loading_indicator.dart';
 import 'package:comics_db_app/ui/navigation/main_navigation.dart';
-import 'package:comics_db_app/ui/widgets/loader_widget/loader_view_model.dart';
+import 'package:comics_db_app/ui/widgets/loader_widget/loader_view_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
